src/student_vladislav_deineka/lesson_5/task_5/ArrayUtil.java
package student_vladislav_deineka.lesson_5.task_5;

import java.util.Arrays;
import java.util.Random;

class ArrayUtil {
    int[] createArray(int arrayLength) {
        return new int[arrayLength];
    }

    public void fillArrayWithRandomNumbers(int[] array) {
        Random random = new Random();
        for (int i = 0; i < array.length; i++) {
            array[i] = random.nextInt();
        }
    }

    public void printArrayToConsole(int[] array) {
        System.out.println(Arrays.toString(array));
    }

    public int findMinNumber(int[] array) {
        int minNumber = array[0];

        for (int i = 0; i < array.length; i++) {
            if (array[i] < minNumber) {
                minNumber = array[i];
            }
        }
        return minNumber;

    }

    public int findMaxNumber(int[] array) {
        int maxNumber = array[4];

        for (int i = 0; i < array.length; i++) {
            if (array[i] > maxNumber) {
                maxNumber = array[i];
            }

        }
        return maxNumber;
    }
}